# docs/design/2026-02-21-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Bid Evaluation Assistant API
  version: v2026.02.21-r3
  description: |
    Gate C OpenAPI contract subset for async job submission, parse/resume/cancel,
    job status query, DLQ operations, citation source lookup, retrieval query/preview,
    and evaluation report lookup.
servers:
  - url: /api/v1
security:
  - BearerAuth: []
paths:
  /documents/upload:
    post:
      operationId: uploadDocument
      summary: Upload a document and enqueue parse/index job
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [project_id, supplier_id, doc_type, file]
              properties:
                project_id:
                  type: string
                supplier_id:
                  type: string
                doc_type:
                  type: string
                  enum: [tender, bid, attachment]
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadAcceptedResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_failed:
                  value:
                    success: false
                    error:
                      code: REQ_VALIDATION_FAILED
                      message: invalid payload
                      retryable: false
                      class: validation
                    meta:
                      trace_id: trace_xxx
                      request_id: req_xxx
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                tenant_scope_violation:
                  value:
                    success: false
                    error:
                      code: TENANT_SCOPE_VIOLATION
                      message: tenant mismatch
                      retryable: false
                      class: security_sensitive
                    meta:
                      trace_id: trace_xxx
                      request_id: req_xxx
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                idempotency_conflict:
                  value:
                    success: false
                    error:
                      code: IDEMPOTENCY_CONFLICT
                      message: same key with different payload
                      retryable: false
                      class: validation
                    meta:
                      trace_id: trace_xxx
                      request_id: req_xxx
  /evaluations:
    post:
      operationId: createEvaluation
      summary: Create an evaluation and enqueue async workflow job
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvaluationRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationAcceptedResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /evaluations/{evaluation_id}/report:
    get:
      operationId: getEvaluationReport
      summary: Get evaluation scoring report with citation coverage fields
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EvaluationId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationReportResponse'
        '404':
          description: Evaluation report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /retrieval/query:
    post:
      operationId: retrievalQuery
      summary: Query retrieval candidates with mode selection and metadata filtering
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrievalQueryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalQueryResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /retrieval/preview:
    post:
      operationId: retrievalPreview
      summary: Preview retrieval evidence without scoring flow
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrievalQueryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalPreviewResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{document_id}/parse:
    post:
      operationId: parseDocument
      summary: Enqueue parse/index job for an uploaded document
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseAcceptedResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{document_id}:
    get:
      operationId: getDocument
      summary: Get document metadata and status
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{document_id}/chunks:
    get:
      operationId: listDocumentChunks
      summary: List parsed chunks for a document
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentChunkListResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /jobs/{job_id}:
    get:
      operationId: getJobStatus
      summary: Query async job status
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant scope violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs:
    get:
      operationId: listJobs
      summary: List jobs with optional filters and cursor pagination
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
      description: Required for all write endpoints
...
    EvaluationId:
      name: evaluation_id
      in: path
      required: true
      schema:
        type: string
...
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    UploadAcceptedResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/UploadAcceptedData'
        meta:
          $ref: '#/components/schemas/Meta'
...
    EvaluationAcceptedResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/EvaluationAcceptedData'
        meta:
          $ref: '#/components/schemas/Meta'
...